Make a public 'OpenAPI' module for PVP interface
