--- openapi/src/OpenAPI/Class.hs
@@ -1,14 +1,0 @@
-module OpenAPI.Class
-  ( -- * Main schema class
-  ToOpenAPISchema(..)
-  -- * Generic schema class
-  , GToOpenAPI(..)
-  -- ** Generic class configuration
-  , GenericSchemaOptions(..)
-  , SumEncoding(..)
-  , defaultSchemaOptions
-  -- ** Entrypoint function for schema deriving
-  , genericToSchema
-  ) where
-
-import OpenAPI.Internal.Class
